+++ frontend/src/App.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { Box, Flex } from '@chakra-ui/react';
 import SideBar from './Components/SideBar.jsx';
-import Chat from './Components/NewChat.jsx';
+import Chat from './Components/Chat.jsx';
 import Agents from './Components/Agents.jsx';
 import SearchQuery from './Components/SearchQuery.jsx';
 import Header from './Components/Header.jsx';
@@ -17,30 +17,75 @@ const App = () => {
 
     // Fetching From Backend
     try {
-        const response = await axios.post('http://127.0.0.1:5000/chat', {
-            message: message,
-        });
+      const response = await fetch('http://localhost:5000/chat', {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({ message: userInput }),
+      });
 
-        const agentReply = response.data.reply;
-        const sources = response.data.sources;
+      const data = await response.json();
 
-        if (typeof agentReply !== 'string') {
-            throw new Error("Agent reply is not a string.");
+      if (data.error) {
+        setMessages(prevMessages => [
+          ...prevMessages,
+          { type: 'error', text: data.error }
+        ]);
+      } else {
+        // the agent response to the chat
+        setMessages(prevMessages => [
+          ...prevMessages,
+          { type: 'agent', text: data.message }
+        ]);
+
+        // Optionally handle courses and faculty lists
+        if (data.faculty.length > 0) {
+          // Convert faculty list to table format
+          const facultyTable = (
+            <table>
+              <thead>
+                <tr>
+                  <th>Faculty Name</th>
+
+                  <th>Gender</th>
+                  &nbsp;
+                  &nbsp;
+                  <th>Department</th>
+                </tr>
+              </thead>
+              <tbody>
+                {data.faculty.map((faculty, index) => (
+                  <tr key={index}>
+                    <td>{faculty.name}</td>
+                    <td>{faculty.gender === 'Male' ? 'Male' : 'Female'}</td>
+                    &nbsp;
+                    &nbsp;
+                    &nbsp;
+                    <td>{faculty.department}</td>
+                  </tr>
+                ))}
+              </tbody>
+            </table>
+          );
+
+          // Add the table to the chat
+          setMessages(prevMessages => [
+            ...prevMessages,
+            { type: 'agent', text: facultyTable }
+          ]);
         }
-
-        setMessages(prevMessages => [
-            ...prevMessages,
-            { text: agentReply, type: 'agent',sources }
-        ]);
-
+      }
     } catch (error) {
-        console.error("Error sending message:", error);
-        setMessages(prevMessages => [
-            ...prevMessages,
-            { text: 'Sorry, something went wrong. Please try again.', type: 'agent' }
-        ]);
+      setMessages(prevMessages => [
+        ...prevMessages,
+        { type: 'error', text: 'Server error. Try again later.' }
+      ]);
+    // } finally {
+    //   // setIsLoading(false);
+    //   setUserInput('');
     }
-};
+  };
   // Handler function to handle agent button clicks
   const handleAgentSelect = (message) => {
     setMessages(prevMessages => [...prevMessages, { text: message, type: 'user' }]);
@@ -55,12 +100,12 @@ const App = () => {
   return (
     <Flex direction="column" h="100vh">
       <div className="fixed  top-0 left-0 w-full">
-      <Header />
+        <Header />
       </div>
       <Flex flex="1" marginTop="4rem">
         <SideBar />
         <Agents onSelectAgent={handleAgentSelect} />
-        <Completion/>
+        <Completion />
         <Chat messages={messages} />
       </Flex>
       <SearchQuery onSend={handleSendMessage} />
