+++ frontend/src/Components/Header.jsx
@@ -1,19 +1,19 @@
-import { Avatar, Text, Stack, Flex } from "@chakra-ui/react";
+import { Avatar, Text, Stack, Flex,Box } from "@chakra-ui/react";
 import logo from '../assets/logo.png';
 
 function Header() {
     return (
-        <div className="m-2">
+        <Box>
             <Flex justify="space-between" align="center">
                 <Flex align="center">
-                    <Avatar src={logo} size='md' boxSize='3rem' />
+                    <Avatar src={logo} size='md' boxSize='3rem' margin='1rem' />
                     <Text fontWeight={600} fontSize={20} ml={2}>AutoGen Bot</Text>
                 </Flex>
                 <Flex cursor='pointer'>
-                    <Avatar size='sm' name="Something New" bg="blue.200" />
+                    <Avatar size='sm' name="Something New" bg="blue.200" margin='1rem' />
                 </Flex>
             </Flex>
-        </div>
+        </Box>
     );
 }
